# docs/swagger.yaml
basePath: /
definitions:
...
  entity.PromoCodeEntity:
    properties:
      activatedWallets:
        description: List of wallets that have activated the promocode
        items:
          type: string
        type: array
      amount:
        description: Reward amount
        type: number
      code:
        description: Unique code for the promocode
        type: string
      createdAt:
        description: Creation timestamp
        type: string
      expiresAt:
        description: Expiration date
        type: string
      id:
        type: string
      maxActivations:
        description: Maximum activations allowed
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/entity.PromoCodeStatus'
        description: Current status
      tokenType:
        description: Reward type
        type: string
      updatedAt:
        description: Last update timestamp
        type: string
      usedActivations:
        description: Number of times the promocode has been used
        type: integer
    type: object
  entity.PromoCodeStatus:
    enum:
    - active
    - expired
    - depleted
    type: string
    x-enum-comments:
      Active: Promocode is active and can be used
      Depleted: Promocode has reached max activations
      Expired: Promocode has expired
    x-enum-varnames:
    - Active
    - Expired
    - Depleted
...
  odm_entities.UserEntity:
    properties:
      created_at:
        type: string
      cubes:
        type: integer
      dfc_balance:
        type: number
      first_name:
        description: Новое поле FirstName
        type: string
      id:
        type: string
      language:
        type: string
      m5_balance:
        type: number
      name:
        type: string
      points:
        type: number
      referral_code:
        type: string
      referral_earnings:
        additionalProperties:
          type: number
        type: object
      referred_by:
        type: string
      tgid:
        type: string
      ton_balance:
        type: number
      updated_at:
        type: string
      wallet:
        type: string
    type: object
...
host: api.m5dice.com
info:
  contact: {}
  description: API for managing users and games in TG-Dice application
  title: TG-Dice API
  version: "1.0"
paths:
...
  /promocodes/{code}:
    get:
      description: Retrieve details of a specific promocode by its code
      parameters:
      - description: Promocode code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Promocode details
          schema:
            $ref: '#/definitions/entity.PromoCodeEntity'
        "404":
          description: Promocode not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get promocode details
      tags:
      - PromoCodes
  /promocodes/activate:
    post:
      consumes:
      - application/json
      description: Activate a promocode by providing the code and user wallet
      parameters:
      - description: Activation request (wallet and code)
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Promocode activated successfully
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Promocode or user not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Activate a promocode
      tags:
      - PromoCodes
  /promocodes/active:
    get:
      description: Retrieve all active promocodes
      produces:
      - application/json
      responses:
        "200":
          description: List of active promocodes
          schema:
            items:
              $ref: '#/definitions/entity.PromoCodeEntity'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List active promocodes
      tags:
      - PromoCodes
  /promocodes/create:
    post:
      consumes:
      - application/json
      description: Create a promocode with details like code, type, amount, and activations
      parameters:
      - description: Promocode details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.PromoCodeEntity'
      produces:
      - application/json
      responses:
        "201":
          description: Promocode created successfully
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new promocode
      tags:
      - PromoCodes
  /promocodes/expire:
    post:
      description: Mark all expired promocodes as expired
      produces:
      - application/json
      responses:
        "200":
          description: Expired promocodes successfully
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Expire outdated promocodes
      tags:
      - PromoCodes
  /referrals/level:
    get:
      consumes:
      - application/json
      description: Возвращает список рефералов на заданном уровне
      parameters:
      - description: Referral code
        in: query
        name: referral_code
        required: true
        type: string
      - description: Level (1, 2, or 3)
        in: query
        name: level
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/odm_entities.UserEntity'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение рефералов по уровню
      tags:
      - referrals
  /referrals/levels:
    get:
      consumes:
      - application/json
      description: Возвращает список рефералов, разделенных по уровням (level1, level2,
        level3)
      parameters:
      - description: Wallet address
        in: query
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение рефералов по уровням
      tags:
      - referrals
  /referrals/total:
    get:
      consumes:
      - application/json
      description: Возвращает общее количество рефералов для данного реферального
        кода
      parameters:
      - description: Referral code
        in: query
        name: referral_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение общего количества рефералов
      tags:
      - referrals
...
schemes:
- https
swagger: "2.0"